@@ -434,15 +434,9 @@ public class Checker extends AbstractAutomaticBean implements MessageDispatcher,
         }
     }
 
-    /**
-     * Performs final setup of the Checker after configuration is complete.
-     *
-     * @noinspection deprecation
-     * @noinspectionreason Disabled until #17646
-     */
     @Override
     protected void finishLocalSetup() throws CheckstyleException {
-        final Locale locale = new Locale(localeLanguage, localeCountry);
+        final Locale locale = Locale.of(localeLanguage, localeCountry);
         LocalizedMessage.setLocale(locale);
 
         if (moduleFactory == null) {
